+++ src/actions/build.rs
@@ -1,35 +1,48 @@
 //! Module for managing and building backends.
 
+use std::path::Path;
+
 use cargo_toml::{Dependency, DependencyDetail, DepsSet};
 
 use crate::{
-    cargo_toml::{odra_dependency, project_name},
+    cargo_toml::odra_dependency,
     command,
+    consts::ODRA_TEMPLATE_GH_RAW_REPO,
     errors::Error,
     log,
-    odra_toml::OdraToml,
+    odra_toml::Contract,
     paths::{self, BuilderPaths},
-    template,
+    project::Project,
+    template::TemplateGenerator,
 };
 
 /// BuildAction configuration.
-pub struct BuildAction {
+pub struct BuildAction<'a> {
     backend: String,
+    contract_name: Option<String>,
     builder_paths: BuilderPaths,
-    odra_toml: OdraToml,
+    project: &'a Project,
+    template_generator: TemplateGenerator,
 }
 
 /// BuildAction implementation.
-impl BuildAction {
+impl<'a> BuildAction<'a> {
     /// Crate a new BuildAction for a given backend.
-    pub fn new(backend: String) -> Self {
+    pub fn new(project: &'a Project, backend: String, contract_name: Option<String>) -> Self {
         BuildAction {
             backend: backend.clone(),
-            builder_paths: BuilderPaths::new(backend),
-            odra_toml: OdraToml::load(),
+            contract_name,
+            builder_paths: BuilderPaths::new(backend, project.project_root.clone()),
+            project,
+            template_generator: TemplateGenerator::new(
+                ODRA_TEMPLATE_GH_RAW_REPO.to_string(),
+                project.project_odra_location(),
+            ),
         }
     }
+}
 
+impl BuildAction<'_> {
     /// Returns the name of the backend.
     /// It is also the name of the Odra's feature.
     pub fn backend_name(&self) -> String {
@@ -40,19 +53,36 @@ impl BuildAction {
     pub fn builder_dependencies(&self) -> DepsSet {
         let mut dependencies = DepsSet::new();
         dependencies.insert(String::from("odra"), self.odra_dependency());
-        dependencies.insert(project_name(), self.project_dependency());
+        self.project.members.iter().for_each(|member| {
+            dependencies.insert(member.name.clone(), self.project_dependency(&member.root));
+        });
         dependencies
     }
 
     /// Main function that runs the whole workflow for a backend.
     pub fn build(&self) {
         self.check_target_requirements();
+        self.validate_contract_name_argument();
         self.prepare_builder();
         self.build_wasm_sources();
         self.build_wasm_files();
         self.format_builder_files();
         self.copy_wasm_files();
         self.optimize_wasm_files();
+    }
+
+    /// Returns list of contract to process.
+    fn contracts(&self) -> Vec<Contract> {
+        let odra_toml = self.project.odra_toml();
+        if let Some(contract_name) = &self.contract_name {
+            odra_toml
+                .contracts
+                .into_iter()
+                .filter(|c| c.name == *contract_name)
+                .collect()
+        } else {
+            odra_toml.contracts.into_iter().collect()
+        }
     }
 
     /// Check if wasm32-unknown-unknown target is installed.
@@ -64,6 +94,21 @@ impl BuildAction {
         }
     }
 
+    /// Check if contract name argument is valid if set.
+    fn validate_contract_name_argument(&self) {
+        if let Some(contract_name) = &self.contract_name {
+            if !self
+                .project
+                .odra_toml()
+                .contracts
+                .iter()
+                .any(|c| c.name == *contract_name)
+            {
+                Error::ContractNotFound(contract_name.clone()).print_and_die();
+            }
+        }
+    }
+
     /// Prepare builder directories and all files.
     fn prepare_builder(&self) {
         log::info(format!(
@@ -72,14 +117,13 @@ impl BuildAction {
             self.builder_paths.root().display()
         ));
 
-        // Prepare directories.
         command::mkdir(self.builder_paths.src());
 
         // Build Cargo.toml
         crate::cargo_toml::builder_cargo_toml(
             &self.builder_paths,
             self.builder_dependencies(),
-            &self.odra_toml,
+            self.contracts(),
         );
 
         // Build files.
@@ -88,10 +132,10 @@ impl BuildAction {
 
     /// Prepare _build.rs files.
     fn create_build_files(&self) {
-        for contract in self.odra_toml.contracts.iter() {
+        for contract in self.contracts() {
             let path = self.builder_paths.wasm_build(&contract.name);
             if !path.exists() {
-                let content = template::wasm_source_builder(
+                let content = self.template_generator.wasm_source_builder(
                     &contract.fqn,
                     &contract.name,
                     &self.backend_name(),
@@ -104,7 +148,7 @@ impl BuildAction {
     /// Prepare _wasm.rs file.
     fn build_wasm_sources(&self) {
         log::info("Generating _wasm.rs files...");
-        for contract in self.odra_toml.contracts.iter() {
+        for contract in self.contracts() {
             command::cargo_build_wasm_sources(self.builder_paths.root(), &contract.name);
         }
     }
@@ -112,7 +156,7 @@ impl BuildAction {
     /// Build _wasm.rs files into .wasm files.
     fn build_wasm_files(&self) {
         log::info("Generating wasm files...");
-        for contract in self.odra_toml.contracts.iter() {
+        for contract in self.contracts() {
             command::cargo_build_wasm_files(self.builder_paths.root(), &contract.name);
         }
     }
@@ -120,10 +164,10 @@ impl BuildAction {
     /// Copy *.wasm files into wasm directory.
     fn copy_wasm_files(&self) {
         log::info("Copying wasm files...");
-        command::mkdir(paths::wasm_dir());
-        for contract in self.odra_toml.contracts.iter() {
-            let source = paths::wasm_path_in_target(&contract.name);
-            let target = paths::wasm_path_in_wasm_dir(&contract.name);
+        command::mkdir(paths::wasm_dir(self.project.project_root()));
+        for contract in self.contracts() {
+            let source = paths::wasm_path_in_target(&contract.name, self.project.project_root());
+            let target = paths::wasm_path_in_wasm_dir(&contract.name, self.project.project_root());
             log::info(format!("Saving {}", target.display()));
             command::cp(source, target);
         }
@@ -131,8 +175,9 @@ impl BuildAction {
 
     /// Run wasm-strip on *.wasm files in wasm directory.
     fn optimize_wasm_files(&self) {
-        for contract in self.odra_toml.contracts.iter() {
-            command::wasm_strip(&contract.name);
+        log::info("Optimizing wasm files...");
+        for contract in self.contracts() {
+            command::wasm_strip(&contract.name, self.project.project_root());
         }
     }
 
@@ -143,28 +188,44 @@ impl BuildAction {
 
     /// Returns Odra dependency tailored for use by builder.
     fn odra_dependency(&self) -> Dependency {
-        match odra_dependency() {
+        let first_member = self.project.members.first().unwrap();
+        match odra_dependency(&first_member.cargo_toml) {
             Dependency::Simple(simple) => Dependency::Detailed(DependencyDetail {
                 version: Some(simple),
                 ..Default::default()
             }),
             Dependency::Detailed(mut odra_details) => {
                 odra_details.features = vec![self.backend_name()];
-                odra_details.default_features = Some(false);
+                odra_details.default_features = false;
                 if odra_details.path.is_some() {
-                    odra_details.path = Some(format!("../{}", odra_details.path.unwrap()));
+                    if self.project.is_workspace() {
+                        odra_details.path = Some(odra_details.path.unwrap());
+                    } else {
+                        odra_details.path = Some(format!("../{}", odra_details.path.unwrap()));
+                    }
                 }
                 Dependency::Detailed(odra_details)
+            }
+            Dependency::Inherited(_) => {
+                Error::NotImplemented("Inherited dependencies are not supported yet.".to_string())
+                    .print_and_die();
             }
         }
     }
 
-    /// Returns project dependency with specific feature enabled.
-    fn project_dependency(&self) -> Dependency {
+    /// Returns project dependency with specific backend feature enabled.
+    fn project_dependency(&self, location: &Path) -> Dependency {
         Dependency::Detailed(DependencyDetail {
-            path: Some("..".to_string()),
+            path: Some(
+                location
+                    .to_path_buf()
+                    .into_os_string()
+                    .to_str()
+                    .unwrap()
+                    .to_string(),
+            ),
             features: vec![self.backend_name()],
-            default_features: Some(false),
+            default_features: false,
             ..Default::default()
         })
     }