+++ src/actions/test.rs
@@ -1,19 +1,21 @@
 //! Module responsible for running contracts tests.
 
 use super::build::BuildAction;
-use crate::{command, paths};
+use crate::{command, log, project::Project};
 
 /// TestAction configuration.
-pub struct TestAction {
+pub struct TestAction<'a> {
+    project: &'a Project,
     backend: Option<String>,
     passthrough_args: Vec<String>,
     skip_build: bool,
 }
 
 /// TestAction implementation.
-impl TestAction {
+impl<'a> TestAction<'a> {
     /// Creates a TestAction struct.
     pub fn new(
+        project: &Project,
         backend: Option<String>,
         passthrough_args: Vec<String>,
         skip_build: bool,
@@ -22,9 +24,12 @@ impl TestAction {
             backend,
             passthrough_args,
             skip_build,
+            project,
         }
     }
+}
 
+impl TestAction<'_> {
     /// Runs a test suite.
     pub fn test(&self) {
         if self.backend.is_none() {
@@ -39,13 +44,15 @@ impl TestAction {
 
     /// Test code against MockVM.
     fn test_mock_vm(&self) {
-        command::cargo_test_mock_vm(paths::project_dir(), self.get_passthrough_args());
+        log::info("Testing against MockVM ...");
+        command::cargo_test_mock_vm(self.project.project_root(), self.get_passthrough_args());
     }
 
     /// Test specific backend.
     fn test_backend(&self) {
+        log::info(format!("Testing backend: {}...", self.backend_name()));
         command::cargo_test_backend(
-            paths::project_dir(),
+            self.project.project_root(),
             self.backend_name(),
             self.get_passthrough_args(),
         );
@@ -63,6 +70,7 @@ impl TestAction {
 
     /// Build *.wasm files before testing.
     fn build_wasm_files(&self) {
-        BuildAction::new(String::from(self.backend_name())).build();
+        BuildAction::new(self.project, String::from(self.backend_name()), None).build();
+        log::info("Building finished.")
     }
 }